make uniform and normal use the same dtype and device

+++ eagerpy/tensor/tensorflow.py
@@ -104,11 +104,15 @@ class TensorFlowTensor(AbstractTensor):
 
     @wrapout
     def uniform(self, shape, low=0.0, high=1.0):
-        return self.backend.random.uniform(shape, minval=low, maxval=high)
+        return self.backend.random.uniform(
+            shape, minval=low, maxval=high, dtype=self.tensor.dtype
+        )
 
     @wrapout
     def normal(self, shape, mean=0.0, stddev=1.0):
-        return self.backend.random.normal(shape, mean=mean, stddev=stddev)
+        return self.backend.random.normal(
+            shape, mean=mean, stddev=stddev, dtype=self.tensor.dtype
+        )
 
     @wrapout
     def ones(self, shape):
